Split charging data job; rtns, mod logs and other
https://eaflood.atlassian.net/browse/WATER-4654

> Part of the work to migrate management of return requirements from NALD to WRLS

For reasons detailed in [Add return requirements clean up step to import](https://github.com/DEFRA/water-abstraction-import/pull/1017) we need to add a new step into the existing 'charging-data' job. But the job itself is already a bit of a mish-mash.

- it is called 'charging-data' but doesn't do that any more
- it brings in fixed reference data, return versions and mod logs all in one go

We know this will soon be ditched, but it is a right old mess as it is. We could be stopping some of the stuff right now, and split out the rest to make it a lot clearer as to what bit is doing what. This wouldn't be much more effort than trying to slot a new step into the existing mess.

So, this change we perform some much needed tidying up of the 'charging-data job'. This should make the change in PR #1017 much easier!