Stop scheduling charging data
What remains in the charging data job is just populating fixed reference data. There is no need to do this as this information never changes.

We leave the job just in case we need to trigger it. But we stop updating these tables on a nightly basis for no reason.