Batch up the work
When we looked at the timings for the existing process, we saw the NALD import took approximately 3.5 hours to process all licences.

Our initial working version was going to take 5 hours (average 4 minutes per 1000) based on the timings we now show (though this was based on my local machine).

Removing the repository and batching the work in sets of 10, I got the time down to under 2 hours (average 1.5 minutes per 1000).