Move all reference data imports to a single process
We shouldn't be importing this data on a nightly basis. We have functionality that depends on this reference so we should be in control of when and how it gets populated.

But just in case we move all the old reference data import jobs to a new 'reference' process which will only ever be manually triggered from here on.