Working version of licence import with error handling
Using `Promise.allSettled` we can deal with individual licences that fail to import and not break the whole import.

Having left the import to run (33 minutes) we dug into the errors we were receiving. One was failing to convert a date. Another was a licence with no non-draft versions. The rest were from not filtering 'draft' status licence versions from the queries for purposes and conditions.