Remove system jobs from licence-import
https://eaflood.atlassian.net/browse/WATER-4670

We've been making changes to the import service to support our work migrating return version management to WRLS, and replacing the current licence import with [one ready for ReSP](https://eaflood.atlassian.net/browse/WATER-4535).

We've also attempted to clarify what is actually happening during the various 'jobs' and to arrange them in our environments in a more logical order.

But the latest version that shipped is failing to complete. We believe the issues are:

- Each 'job' is taking longer to complete than expected
- Based on current timings, the 'jobs' are overlapping
- Our attempts to connect the new ReSP import with the existing one are adding to the burden
- Based on investigating the logs, we think downstream jobs are no longer being triggered

We're [working on simplifying the project](https://github.com/DEFRA/water-abstraction-import/pull/1023) and removing those elements we think decrease performance, for example, the use of [pg-boss](https://github.com/timgit/pg-boss).

But in the meantime we need to get the import back up and running. This change removes the extra jobs we put in for connecting to the [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) till we can come up with a better alternative. We also update the default schedule to allow each part of the import more time to complete to avoid overlapping.