Import means of abstraction for points
https://eaflood.atlassian.net/browse/WATER-4645

> Part of the work to migrate return versions from NALD to WRLS

In [Implement new 'points' job](https://github.com/DEFRA/water-abstraction-import/pull/1028) we amended the import service to include a new 'points' job. Having realised that the way were intending to store point information was insufficient, we made enough changes to our points schema to warrant its own job to manage importing the NALD data.

We thought we'd captured everything 'point' related in the change but have since spotted 'Means of abstraction'. This is linked to a point in NALD and gives how water will be abstracted, for example, "Surface Mounted Pump (Fixed)". We are required to display this in the UI. Having [added the new column](https://github.com/DEFRA/water-abstraction-service/pull/2642), this ensures it gets populated.