Refactor `licence-document` clean job https://eaflood.atlassian.net/browse/WATER-4726
Prior to carrying out the work in ticket 4726 we are going to refactor the `licence-import.delete-removed-documents` job and make it a more generic `clean` job as we did in PR https://github.com/DEFRA/water-abstraction-import/pull/1017.

Once this refactoring has been done we will add some new queries to the `clean` job in a separate PR to remove imported data related to the licence which no longer exists in NALD, but do exist in our system.