Fix broken crm_v2.documents constraint (#1033)
https://eaflood.atlassian.net/browse/WATER-4708

In a [recent change to water-abstraction-tactical-crm](https://github.com/DEFRA/water-abstraction-tactical-crm/pull/1053), we altered the constraint on the `crm_v2.documents` table to remove fields that are always set to the same values. It helped simplify our work to look at migrating the import process to [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system).

However, we failed to realise that the import refers to those fields in a query it uses to import NALD licences to the `crm_v2.documents` table. This means our change has broken this element of the NALD import.

This change updates the query to match the constraint's current signature. Once deployed, the import will automatically pick up any NALD licences that need updates applying in `crm_v2.documents`.