Cleanup `licence_version_purposes` table (#1036)
https://eaflood.atlassian.net/browse/WATER-4732

Update the `licence-import` process to identify records in the `licence_version_purposes` table that no longer exist in NALD and remove those records. Any child records must also be removed before the parent record.

This functionality will temporarily be put behind a feature flag `CLEAN_LICENCE_IMPORTS`, which will need to be set to `true` for the new code to run.