Cleanup `licence_versions` table https://eaflood.atlassian.net/browse/WATER-4733
Update the `licence-import` process to identify records in the `licence_versions` table that no longer exist in NALD and remove those records.

The only child record that hangs off of this table appears to the `licence_version_purposes` which has already been removed in this PR https://github.com/DEFRA/water-abstraction-import/pull/1036

This functionality will temporarily be put behind a feature flag `CLEAN_LICENCE_IMPORTS`, which will need to be set to `true` for the new code to run.