Cleanup `licences` table (#1038)
https://eaflood.atlassian.net/browse/WATER-4734

Update the `licence-import` process to identify records in the `licences` table that no longer exist in NALD and remove those records. If the licence has been involved in a bill run or has a return version, it should not be removed.

All child records also need to be removed. The `licence_versions` and its child records have already been removed in previous PRs https://github.com/DEFRA/water-abstraction-import/pull/1037

This functionality will temporarily be put behind a feature flag `CLEAN_LICENCE_IMPORTS`, which will need to be set to `true` for the new code to run.