Delete "licences" that are not in NALD https://eaflood.atlassian.net/browse/WATER-4734
Update the `licence-import` process to identify records in the `licences` table that no longer exist in NALD and remove those records.

If the licence has been involved in a bill run, has a return version, or a purpose condition associated with a monitoring station, it should not be removed. All child records also need to be removed.

This functionality will temporarily be put behind a feature flag `CLEAN_LICENCE_IMPORTS`, which will need to be set to `true` for the new code to run.