Deal with deleted lics in rtn ver import & clean
https://eaflood.atlassian.net/browse/WATER-4654

> Part of the work to migrate management of return versions from NALD to WRLS

As part of migrating management of return versions to WRLS we have been making a series of changes to [water-abstraction-import](https://github.com/DEFRA/water-abstraction-import) to ensure our data is as 'clean' as possible. Users in NALD are able to delete records but until we started this work, they were never deleted on the WRLS side.

We moved everything related to return versions into its own job, then updated the import side to ensure the NALD data was transformed to a state we can work with, and parts that were previously not included such as points are now imported. After that we then added a clean step to ensure return information that no longer existed in NALD was also deleted in WRLS.

This work has fed the next data clean up task: cleaning up the licence information. One of the things that will block us deleting a licence from WRLS (that has been deleted in NALD) is if it has a 'completed' return log. This is external user submitted data so we can't delete it, nor can we delete the licence record it is linked to, even if it has been deleted in NALD.

We made an assumption that when a licence is deleted in NALD _all_ child records are also deleted. So, when it comes to deleting the licence, we can assume if a `water.return_version` record still exists in WRLS it is because we blocked deletion due to a 'completed' return log.

It turns out that core assumption is wrong. There are examples of licences that have been deleted from NALD, but their child records still exist, including return versions. This means our import is importing data for deleted licences, and our clean is never removing it because the record _is in_ NALD!

This change updates our return versions job to cater for what we now know.

- **import** - exclude return information where the licence record no longer exists
- **clean** - include return information where the child record might exist, but the parent doesn't