Cleanup licence data removed from NALD https://eaflood.atlassian.net/browse/WATER-4662
Following a requirements analysis meeting on the 21st November we need to make some updates to the queries we have written to remove licence data from WRLS that no longer exists in NALD.

A PR that deals with the scenario where the licence record has been removed from NALD has already been completed in PR https://github.com/DEFRA/water-abstraction-import/pull/1044

This PR will be focussing on the child records of the licence that may have been removed from NALD whilst the licence still remains. The exception to this would be a licence that has been removed from NALD but has not been removed from WRLS as it has been involved in a bill run, has a return version, or is linked to an external entity.

The individual tickets covered by this PR are:
- https://eaflood.atlassian.net/browse/WATER-4751