Update `cleanNaldLicenceVersionPurposePoints` query At this point in the process the only monitoring stations remaining are ones which should not be removed. As they can't be removed that means that the purpose will also not be removed.
Because the purpose cannot be removed (even if it has been removed from NALD) it will need at least 1 point to be a valid record. Therefore the point is not deleted. This is because if there were 2 points on the purpose both could also have been deleted from NALD if the purpose was removed. This would make it pretty much impossible to determine which point may need to be kept if there are more than one. Therefore no point gets deleted.

The same applies to the purpose if its associated licence cannot be deleted for reasons.