Add plugin test for Import Points job
We have a policy of limiting our effort in the legacy code as the goal is to ditch it anyway. But the jump from 5 to 7 in the tests looked odd. So, adding a missing block for Import Points to the plugin test so it doesn't cause confusion in the future.