Move trigger for licence end dates in lic. import (#1050)
https://eaflood.atlassian.net/browse/WATER-4546

> Part of the work to migrate management of return versions to WRLS from NALD

In [Trigger new licence end date change endpoints](https://github.com/DEFRA/water-abstraction-import/pull/1049), we added new steps to the NALD and Licence import to allow us to trigger functionality in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) to trigger functionality linked to changes to licence end dates.

It needs two steps because in the **NALD import**, we need to see the difference between NALD and WRLS end dates _before_ the WRLS licence record is updated. Then, in **Licence import**, we can process the changes _after_ the licence has been updated with the new end date.

Sorry, I did say _after_ didn't I?! 🤦 

We realised in testing that we'd placed the second step in **Licence import** before it updates the licence. This means features that depend on knowing what the change date was and that the licence has been updated with the new end date are failing.

This fix moves the step in **Licence import** to _after_ the steps that update the licence records.