Add custom serializers to improve app logging (#1051)
We've completed a [number of changes to our logging](https://github.com/DEFRA/water-abstraction-team/issues?q=is%3Aissue%20state%3Aclosed%20log) in the time we have been on the project.

Generally, we 'go to the logs' when an issue has been reported in `production`, and we need to figure out what has happened.

Therefore, we constantly balance, ensuring we have all the necessary information while excluding noise.

We recently [made a change to water-abstraction-system](https://github.com/DEFRA/water-abstraction-system/pull/1655) to reduce the noise in our logs via custom serializers (check it out for more details).

This change makes the same change to the logging in this app.