Add helper function to main db module
The licence import has its own library that does this. It means the functions that query the DB no longer have to extract the rows from the result object returned.

To us, it makes more sense to put the helper function on `db.js` rather than hidden away in a jobs `lib`.