Refactor new return-logs import job to use DB
https://eaflood.atlassian.net/browse/WATER-4855

> Part of the work to migrate management of return requirements from NALD to WRLS

We've realised NALD has return log data that we don't. Initially, we thought it was just the return lines for pre-2013 return logs. But after some digging, it appears there are not only the return submission lines but also some return logs we don't have that NALD does.

Putting aside the missing return logs, we need to get those missing return submission lines in. We have found that NALD structures the return logs and submission data very differently from WRLS. Our plan to do something 'quick & dirty' with a SQL query is therefore not possible. That means we'll need to code a solution.

We intend to stop all 'return leg' processes in the very near future, just as soon as WRLS takes over from NALD on all things 'returns' related. So, it makes no sense to start from scratch. This project understands how to interpret and create WRLS return logs and submission data based on what NALD holds. We could work with it and make the necessary changes to get this missing data in, including the missing return logs.

The first step of that was [Move import of return logs into their own job](https://github.com/DEFRA/water-abstraction-import/pull/1055).

The next step is to replace all the requests to [water-abstraction-returns](https://github.com/DEFRA/water-abstraction-returns) with straightforward queries against the DB. This unnecessary complication not only makes it slower, but much harder to follow what changes the legacy code is actually making. It also resolves an issue we have with the **water-abstraction-returns** app, which breaks environments due to the amount of logging this interaction generates.