Replace first API query with DB one
The first 'query' the job was making, was for every return requirement for a licence (in NALD), after working out the WRLS `return_id` for the return requirement, checking if the return log already existed.

So, we have 74K licences. Assume 70K of them have return requirements. If you are nice and say each has only 2 return requirements, that is 140K API requests from this service to **water-abstraction-returns** that was being made then logged!