Make return exists private
Realised it was only 'public' and exported on the module for the purposes of testing. We struggle with the legacy code understanding what might be used elsewhere and what is just for use within a module because of this pattern the previous team used.

So, when we are sure (and we have 5 mins) we update the functions to make this clear.