Bring the persist-returns tests inline
Excluding the use of `experiment` and `test` over `describe` and `it`, this is how we structure our test files and we find it makes it much easier to understand what is being reference in the file and why.

We also remove all the unnecessary async declarations we've just realised were in the tests.