Move delete returns in connector to own module
Looks like this never went through the returns API and instead was making direct queries against the DB!

Anyhoo, we move it out of the connector so we can see more clearly what is still hitting the API and what is talking direct to the DB.

We also drop it's deleting of the return versions. That's all handled and unlike all these other legacy jobs, we delete what no longer exists in NALD as part of the process. There is no need for this job to muddy the waters, even if it is only for a feature intended to be used in non-production environments.