Refactor new return-logs import job to use DB (#1056)
https://eaflood.atlassian.net/browse/WATER-4855

> Part of the work to migrate management of return requirements from NALD to WRLS

We've realised NALD has return log data that we don't. Initially, we thought it was just the return lines for pre-2013 return logs. But after some digging, it appears there are not only the return submission lines but also some return logs we don't have that NALD does.

Putting aside the missing return logs, we need to get those missing return submission lines in. We have found that NALD structures the return logs and submission data very differently from WRLS. Our plan to do something 'quick & dirty' with a SQL query is therefore not possible. That means we'll need to code a solution.

We intend to stop all 'return leg' processes in the very near future, just as soon as WRLS takes over from NALD on all things 'returns' related. So, it makes no sense to start from scratch. This project understands how to interpret and create WRLS return logs and submission data based on what NALD holds. We could work with it and make the necessary changes to get this missing data in, including the missing return logs.

The first step of that was [Move import of return logs into their own job](https://github.com/DEFRA/water-abstraction-import/pull/1055).

The next step is to replace all the requests to [water-abstraction-returns](https://github.com/DEFRA/water-abstraction-returns) with straightforward queries against the DB. This unnecessary complication makes it slower and much more challenging to follow the changes the legacy code is making. It also resolves an issue with the **water-abstraction-returns** app, which breaks environments due to the amount of logging this interaction generates.

---

Those requests that support the non-production 'replicate returns' functionality have been left in. We began with an intent to replace them with queries, and then use them to support bringing in the missing returns data.

But after digging into them, we have found the logic they use is faulty. We have found instances of returns where

- 0 quantity lines have been ignored. They should be migrated as this is a customer actively recording that no water was abstracted
- weekly returns where lines are missing, for example, a 2013-04-01 to 2014-03-31 return log will only have lines for the weeks in 2013.

The same logic for displaying NALD data as a WRLS return also exists in [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service) and it does not have these issues.

So, in the next series of changes we'll replace the faulty logic with an engine that not only works, but can bring I our missing data.