Get new replicate engine matching existing returns
The engine we are replacing in here had an approach where it would extract monthly returns data from the `NALD_RET_FORM_LOGS` table, and weekly and daily from `NALD_RET_LINES`.

But we also know it was ignoring `0` value entries in either table, and seemed to be messing up weekly returns. If the return log crossed a year only the first year's weeks would get lines.

The engine we found in **water-abstraction-service** does a better job. `0` value entries are not excluded and it doesn't appear to miss entries. It's issue is the entry dates don't always line up with a WRLS return.

For example, in WRLS a 'week' runs Sunday to Saturday. So, all weekly line entry end dates are Saturdays. But that is not the case in NALD. You can have a weekly return where Friday might be the week end day!

When they generate the weekly line entry, they calculate the start date as per WRLS rules, but they use the date from the NALD line as the end date. This means you get gaps, because their weekly line entry runs Sunday to Friday. Plus, it being the original team, it's needlessly complex. Why did someone feel the need to generate a histogram to work out which null NALD line entries to ignore!?

Ours we believe is both simpler, and accurate. Given a return log generated by the import, work out the WRLS days, week, or month line entries as per our rules. So, no gaps!

Then match NALD lines to these lines, sum them, and assign the total as the line quantity. We know thanks to the **water-abstraction-service** engine we don't have to worry about looking at `NALD_RET_FORM_LOGS` for the monthly data. Plus, no histograms in sight!

We're not persisting anything yet, but we're committing the changes before our next challenge, determining what a valid version looks like. The one in this change comes from the **water-abstraction-service** engine which only has to worry about the legacy view page.