Delete both queues on a manual trigger
When testing we weren't seeing the licence we passed in being processed. Instead, it seemed to be working through a bunch of others. That's when we realised that the we were only deleting the 'queue' queue on a manual trigger. Anything that been queued up in the 'import' queue would continue to be processed.