Update version generation
Looking at all the recent return submissions that have been added by internal users, the `metadata` is set to an empty object. There is nothing in the row metadata that is not already either in the version's other fields or the parent return log's metadata field.

So, we keep things simple and just set `metadata` to an empty object in the return submission.