Replace faulty replicate returns logic (#1057)
https://eaflood.atlassian.net/browse/WATER-4855

> Part of the work to migrate management of return requirements from NALD to WRLS

We've realised NALD has return log data that we don't. Initially, we thought it was just the return lines for pre-2013 return logs. But after some digging, it appears there are not only the return submission lines but also some return logs we don't have that NALD does.

Putting aside the missing return logs, we need to get those missing return submission lines in. We have found that NALD structures the return logs and submission data very differently from WRLS. Our plan to do something 'quick & dirty' with a SQL query is therefore not possible. That means we'll need to code a solution.

We intend to stop all 'return leg' processes in the very near future, just as soon as WRLS takes over from NALD on all things 'returns' related. So, it makes no sense to start from scratch.

Or it didn't when we thought this project understood how to interpret and create WRLS return logs and submission data based on what NALD holds.

Turns out it doesn't!

We have found the replicate returns logic is faulty. There are instances of returns where

- 0 quantity lines have been ignored. They should be migrated as this is a customer actively recording that no water was abstracted
- weekly returns where lines are missing, for example, a 2013-04-01 to 2014-03-31 return log will only have lines for the weeks in 2013

We might have opted to start from scratch if we knew that when we started!

However, no one has complained about its handling of return logs. Plus, we recently learned that when displaying a return that ends before 2018-10-31, the legacy UI will query the `import` schema in the DB directly, even if there is data in the `returns` schema. 😩🤦

After some digging, we found it much more successful at interpreting NALD data than what was built in this project. So, having come this far, the next step is to update the faulty replicate returns engine with the logic from [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service).

Having proven it with the data we have in non-production environments, the next steps will be integrating it as part of the normal import process where we are missing submission data and getting it to look at the missing NALD line data we've obtained. But that's for later! For now, we have a crappy engine to replace 😀💪